move fd_lock_path and path_lock fns from forc-pkg to forc-util
