+++ forc-pkg/Cargo.toml
@@ -12,7 +12,6 @@ repository.workspace = true
 ansi_term = "0.12"
 anyhow = "1"
 cid = "0.10"
-fd-lock = "3.0"
 forc-tracing = { version = "0.46.0", path = "../forc-tracing" }
 forc-util = { version = "0.46.0", path = "../forc-util" }
 fuel-abi-types = "0.1"
