+++ forc-pkg/src/source/ipfs.rs
@@ -62,7 +62,7 @@ impl source::Fetch for Pinned {
             anyhow::bail!("offline fetching for IPFS sources is not supported")
         }
 
-        let mut lock = crate::pkg::path_lock(repo_path)?;
+        let mut lock = forc_util::path_lock(repo_path)?;
         {
             let _guard = lock.write()?;
             if !repo_path.exists() {
@@ -111,7 +111,7 @@ impl source::DepPath for Pinned {
     fn dep_path(&self, name: &str) -> anyhow::Result<source::DependencyPath> {
         let repo_path = pkg_cache_dir(&self.0);
         // Co-ordinate access to the ipfs checkout directory using an advisory file lock.
-        let lock = crate::pkg::path_lock(&repo_path)?;
+        let lock = forc_util::path_lock(&repo_path)?;
         let _guard = lock.read()?;
         let path = manifest::find_within(&repo_path, name)
             .ok_or_else(|| anyhow::anyhow!("failed to find package `{}` in {}", name, self))?;
